Support zero start/end times to create view across entire source
In some cases, like PCAPs or smaller files it can be helpful to
query the entire timerange of the source rather than first determining
the start/end ranges which can be in the far distant past.

These changes will allow passing both start and end as None values
so no TimeFilter gets applied to the created view.